+++ integrations/datadog/webhook_processors/monitor_webhook_processor.py
@@ -1,11 +1,7 @@
-import base64
-from typing import Any
-from initialize_client import init_client
-from port_ocean.context.ocean import ocean
-from port_ocean.core.handlers.port_app_config.models import ResourceConfig
-from port_ocean.core.handlers.webhook.abstract_webhook_processor import (
-    AbstractWebhookProcessor,
+from webhook_processors._abstract_webhook_processor import (
+    _AbstractDatadogWebhookProcessor,
 )
+from port_ocean.core.handlers.port_app_config.models import ResourceConfig
 from port_ocean.core.handlers.webhook.webhook_event import (
     EventPayload,
     WebhookEvent,
@@ -14,7 +10,7 @@ from port_ocean.core.handlers.webhook.webhook_event import (
 from integration import ObjectKind
 
 
-class MonitorWebhookProcessor(AbstractWebhookProcessor):
+class MonitorWebhookProcessor(_AbstractDatadogWebhookProcessor):
     async def should_process_event(self, event: WebhookEvent) -> bool:
         return True
 
@@ -24,32 +20,12 @@ class MonitorWebhookProcessor(AbstractWebhookProcessor):
     async def handle_event(
         self, payload: EventPayload, resource_config: ResourceConfig
     ) -> WebhookEventRawResults:
-        dd_client = init_client()
-        monitor = await dd_client.get_single_monitor(payload["alert_id"])
+        monitor = await self._webhook_client.get_single_monitor(payload["alert_id"])
 
         return WebhookEventRawResults(
             updated_raw_results=[monitor] if monitor else [],
             deleted_raw_results=[],
         )
-
-    async def authenticate(
-        self, payload: EventPayload, headers: dict[str, Any]
-    ) -> bool:
-        authorization = headers.get("authorization")
-        webhook_secret = ocean.integration_config.get("webhook_secret")
-
-        if authorization:
-            try:
-                auth_type, encoded_token = authorization.split(" ", 1)
-                if auth_type.lower() != "basic":
-                    return False
-
-                decoded = base64.b64decode(encoded_token).decode("utf-8")
-                _, token = decoded.split(":", 1)
-                return token == webhook_secret
-            except (ValueError, UnicodeDecodeError):
-                return False
-        return True
 
     async def validate_payload(self, payload: EventPayload) -> bool:
         return "event_type" in payload and "alert_id" in payload
